add navigator
