+++ package.json
@@ -24,13 +24,15 @@
     "react-native-config": "^1.4.6",
     "react-native-css-vh-vw": "^1.0.5",
     "react-native-gesture-handler": "^2.5.0",
+    "react-native-image-picker": "^4.10.0",
     "react-native-nmap": "^0.0.66",
     "react-native-reanimated": "^2.9.1",
     "react-native-recoil-persist": "^0.0.4",
     "react-native-safe-area-context": "^4.3.1",
     "react-native-screens": "^3.15.0",
     "react-native-vector-icons": "^9.2.0",
-    "recoil": "^0.7.4"
+    "recoil": "^0.7.4",
+    "rn-range-slider": "^2.1.1"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
